+++ docs/Changelog.md
@@ -18,12 +18,16 @@
     - allow passing classLoader at `ResourceLoader.load`
 * [#4574](https://github.com/TouK/nussknacker/pull/4574) Improvements: at `KafkaClient` and `RichKafkaConsumer` in kafka-test-utils
 <<<<<<< HEAD
+<<<<<<< HEAD
 * [#4640](https://github.com/TouK/nussknacker/pull/4640) Expand timestamp support to handle more types/formats
 * [#4685](https://github.com/TouK/nussknacker/pull/4685) App API OpenAPI-based documentation (e.g. `https://demo.nussknacker.io/api/docs`)
 * [#4707](https://github.com/TouK/nussknacker/pull/4707) Support for `overrideFrontendAuthenticationStrategy` configuration parameter in OIDC security model - works the same as in OAuth2 case.
 =======
 * [#4287](https://github.com/TouK/nussknacker/pull/4287) Improvements: bump cats effect
 >>>>>>> 34841267d2 (Add docs)
+=======
+* [#4287](https://github.com/TouK/nussknacker/pull/4287) Improvement: bump cats effect to 3.5.1
+>>>>>>> 1be91a59b6 (Upgrade cats and remove using IORuntime.global)
 
 1.11.3 (11 Sep 2023)
 -------------------------